laplasl360
This is a direct source copy of apl360, I will be referencing this for the creation of a new, better, and smaller version of a foundation software with out  a gui, or operating system.  The goal is to to have a user terminal I/O interface, that literally queries the hardware.